more merge fixes fixes
