+++ modules/custom/dkan_api/src/Storage/DrupalNodeDataset.php
@@ -21,7 +21,7 @@ class DrupalNodeDataset implements StorageInterface {
   /**
    * @var Drupal\dkan_api\Storage\ThemeValueReferencer
    */
-  private $themeValueReferencer;
+  protected $themeValueReferencer;
 
   /**
    * Constructor.
