+++ modules/custom/dkan_api/tests/src/Unit/Storage/DrupalNodeDatasetTest.php
@@ -6,6 +6,7 @@ use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\dkan_common\Tests\DkanTestBase;
 use Drupal\dkan_api\Storage\DrupalNodeDataset;
 use Drupal\node\NodeStorageInterface;
+use Drupal\dkan_api\Storage\ThemeValueReferencer;
 
 /**
  * Tests Drupal\dkan_api\Storage\DrupalNodeDataset.
@@ -21,17 +22,21 @@ class DrupalNodeDatasetTest extends DkanTestBase {
    */
   public function testConstruct() {
     $mockEntityTypeManager = $this->createMock(EntityTypeManagerInterface::class);
-
+    $mockThemeValueReferencer = $this->createMock(ThemeValueReferencer::class);
     $mock = $this->getMockBuilder(DrupalNodeDataset::class)
       ->disableOriginalConstructor()
       ->getMock();
 
     // Assert.
-    $mock->__construct($mockEntityTypeManager);
+    $mock->__construct($mockEntityTypeManager, $mockThemeValueReferencer);
 
     $this->assertSame(
             $mockEntityTypeManager,
             $this->readAttribute($mock, 'entityTypeManager')
+    );
+    $this->assertSame(
+            $mockThemeValueReferencer,
+            $this->readAttribute($mock, 'themeValueReferencer')
     );
   }
